+++ app/functions/commands/hearsphoto.ts
@@ -8,6 +8,7 @@
  * @license: MIT License
  *
  */
+import { InlineKeyboard } from "grammy";
 import bot from "@app/core/token";
 import translate from "@translations/translate";
 import db from "@routes/api/database";