+++ CHANGELOG.md
@@ -1,12 +1,22 @@
-# v0.5.0 (November 21, 2021)
+# v0.6.0 (November 24, 2021)
 
--   Porting: from Telegraf to Grammy Framework
+-   Refactor: No Any
+-   Feature: /show - current quiz
+-   Feature: /ping - call to play
+-   Feature: /groups - official groups
+-   Fix: unpin current quiz when the user pick the master role
+-   Fix: remove up/down vote from previously quiz
+-   Fix: default language from telegram api
 
 <!-- all-shields/sponsors-badges:START -->
 
 [![Donate Paypal](https://img.shields.io/badge/donate-paypal-005EA6.svg?style=for-the-badge&logo=paypal)](https://www.paypal.me/ptkdev) [![Donate Ko-Fi](https://img.shields.io/badge/donate-ko--fi-29abe0.svg?style=for-the-badge&logo=ko-fi)](https://ko-fi.com/ptkdev) [![Donate Github Sponsors](https://img.shields.io/badge/donate-sponsors-ea4aaa.svg?style=for-the-badge&logo=github)](https://github.com/sponsors/ptkdev) [![Donate Patreon](https://img.shields.io/badge/donate-patreon-F87668.svg?style=for-the-badge&logo=patreon)](https://www.patreon.com/join/ptkdev) [![Donate Bitcoin](https://img.shields.io/badge/BTC-35jQmZCy4nsxoMM3QPFrnZePDVhdKaHMRH-E38B29.svg?style=flat-square&logo=bitcoin)](https://ptk.dev/img/icons/menu/bitcoin_wallet.png) [![Donate Ethereum](https://img.shields.io/badge/ETH-0x8b8171661bEb032828e82baBb0B5B98Ba8fBEBFc-4E8EE9.svg?style=flat-square&logo=ethereum)](https://ptk.dev/img/icons/menu/ethereum_wallet.png)
 
 <!-- all-shields/sponsors-badges:END -->
+
+# v0.5.0 (November 21, 2021)
+
+-   Porting: from Telegraf to Grammy Framework
 
 # v0.4.0 (November 07, 2021)